weather-forecast/ecrad-3d-correction/cnn+rnn/requirements.txt
black==24.10.0
climetlab==0.24.0
climetlab-maelstrom-radiation==0.7.5
lightning==2.4.0
lightning-utilities==0.11.8
numpy==1.26.3
nvidia-dali-cuda120==1.42.0
nvidia-ml-py==12.535.161
nvidia-pyindex==1.0.9
nvitop==1.3.2
omegaconf==2.3.0
pytorch-lightning==2.4.0
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.14.0
torchaudio==2.5.0+cu121
torchexplorer==1.1.2
torchmetrics==1.5.1
torchvision==0.20.0+cu121
transformers==4.46.2
xarray[complete]==2024.10.0

